+++ packages/extension-base/src/services/balance-service/transfer/token.ts
@@ -9,6 +9,7 @@ import { estimateTonTxFee } from '@subwallet/extension-base/services/balance-ser
 import { _TRANSFER_CHAIN_GROUP } from '@subwallet/extension-base/services/chain-service/constants';
 import { _EvmApi, _SubstrateApi, _TonApi } from '@subwallet/extension-base/services/chain-service/types';
 import { _getContractAddressOfToken, _getTokenOnChainAssetId, _getTokenOnChainInfo, _getXcmAssetMultilocation, _isBridgedToken, _isChainEvmCompatible, _isChainTonCompatible, _isGigaToken, _isNativeToken, _isTokenGearSmartContract, _isTokenTransferredByEvm, _isTokenTransferredByTon, _isTokenWasmSmartContract } from '@subwallet/extension-base/services/chain-service/utils';
+import { TaoStakeInfo } from '@subwallet/extension-base/services/earning-service/handlers/native-staking/tao';
 import { calculateGasFeeParams } from '@subwallet/extension-base/services/fee-service/utils';
 import { createLogger } from '@subwallet/extension-base/utils/logger';
 import { combineEthFee, getGRC20ContractPromise, getVFTContractPromise } from '@subwallet/extension-base/utils';
@@ -126,6 +127,15 @@ export const createSubstrateExtrinsic = async ({ from, networkKey, substrateApi,
     }
   } else if (_TRANSFER_CHAIN_GROUP.truth.includes(networkKey)) {
     transfer = api.tx.assetManager.transfer(to, _getTokenOnChainInfo(tokenInfo), value);
+  } else if (_TRANSFER_CHAIN_GROUP.bittensor.includes(networkKey) && tokenInfo.metadata?.netuid) {
+    const { accountHotKey, isEnableTransferSubnet } = await getAccountNetuidTokenInfo(from, tokenInfo, substrateApi);
+    const tokenNetuid = tokenInfo.metadata?.netuid;
+
+    if (isEnableTransferSubnet && accountHotKey) {
+      transfer = api.tx.subtensorModule.transferStake(from, accountHotKey, tokenNetuid, tokenNetuid, value);
+    } else {
+      return [null, value];
+    }
   }
 
   return [transfer, transferAmount || value];
@@ -189,4 +199,23 @@ export const getTransferMockTxFee = async (address: string, chainInfo: _ChainInf
 
     return new BigN(0);
   }
+};
+
+export const getAccountNetuidTokenInfo = async (address: string, tokenInfo: _ChainAsset, substrateApi: _SubstrateApi): Promise<{isEnableTransferSubnet: boolean, accountHotKey: string | undefined}> => {
+  const tokenNetuid = tokenInfo.metadata?.netuid;
+
+  if (!tokenNetuid) {
+    return { isEnableTransferSubnet: false, accountHotKey: undefined };
+  }
+
+  const [isTransferToggle, rawStakeInfoForColdKey] = await Promise.all([
+    substrateApi.api.query.subtensorModule.transferToggle(tokenNetuid),
+    substrateApi.api.call.stakeInfoRuntimeApi.getStakeInfoForColdkey(address)
+  ]);
+
+  const isEnableTransferSubnet = isTransferToggle.toPrimitive() as boolean;
+  const taoStakeInfos = rawStakeInfoForColdKey.toPrimitive() as unknown as TaoStakeInfo[];
+  const accountHotKey = taoStakeInfos.find((value) => value.netuid === tokenNetuid);
+
+  return { isEnableTransferSubnet, accountHotKey: accountHotKey?.hotkey };
 };