+++ packages/extension-base/src/services/event-service/index.ts
@@ -43,6 +43,8 @@ export class EventService extends EventEmitter<EventRegistry> {
   public readonly waitEarningReady: Promise<boolean>;
   public readonly waitLedgerReady: Promise<boolean>;
 
+  public readonly waitOpenGovReady: Promise<boolean>;
+
   constructor () {
     super();
     this.timeoutId = null;
@@ -67,6 +69,8 @@ export class EventService extends EventEmitter<EventRegistry> {
     this.waitBuyServiceReady = this.generateWaitPromise('buy.services.ready');
     this.waitEarningReady = this.generateWaitPromise('earning.ready');
     this.waitLedgerReady = this.generateWaitPromise('ledger.ready');
+
+    this.waitOpenGovReady = this.generateWaitPromise('open-gov.ready');
   }
 
   private generateWaitPromise<T extends EventType> (eventType: T): Promise<boolean> {