+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # CHANGELOG
 
+## 1.3.71
+Build date: Dec 29, 2025.
+
+Features & Updates:
+- Improve token enabling (Round 2) (#4247)
+- Update libs for SubWallet Extensions (#4808)
+- Support Import from Trust Wallet to SubWallet (#4762)
+
+
+## 1.3.70
+Build date: Dec 11, 2025.
+
+Features & Updates:
+- Support OpenGov (Phase 1) (#4678)
+
+
+## 1.3.69
+Build date: Dec 08, 2025.
+
+Features & Updates:
+- Update chain-list stable v0.2.122 (#4827)
+- Turn off "Advanced phishing detection" feature (#4891)
+
+
 ## 1.3.68
 Build date: Dec 03, 2025.